+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `FE` Removed redundant clock icon from HUD.
+
 ## [1.0.0] - 2022-04-21
 
 ### Added